requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
absl-py==0.11.0
    # via
    #   tensorboard
    #   tensorflow
asgiref==3.3.1
    # via django
astunparse==1.6.3
    # via tensorflow
cachetools==4.2.1
    # via google-auth
certifi==2020.12.5
    # via requests
chardet==4.0.0
    # via requests
click==7.1.2
    # via uvicorn
django-cors-headers==3.7.0
    # via -r requirements.in
django==3.1.7
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   djangorestframework
djangorestframework==3.12.2
    # via -r requirements.in
fastapi==0.63.0
    # via -r requirements.in
flatbuffers==1.12
    # via tensorflow
gast==0.3.3
    # via tensorflow
google-auth-oauthlib==0.4.3
    # via tensorboard
google-auth==1.27.1
    # via
    #   google-auth-oauthlib
    #   tensorboard
google-pasta==0.2.0
    # via tensorflow
grpcio==1.32.0
    # via
    #   tensorboard
    #   tensorflow
gunicorn==20.0.4
    # via -r requirements.in
h11==0.12.0
    # via uvicorn
h5py==2.10.0
    # via tensorflow
httptools==0.1.1
    # via uvicorn
idna==2.10
    # via requests
keras-preprocessing==1.1.2
    # via tensorflow
markdown==3.3.4
    # via tensorboard
numpy==1.19.5
    # via
    #   h5py
    #   keras-preprocessing
    #   opt-einsum
    #   tensorboard
    #   tensorflow
oauthlib==3.1.0
    # via requests-oauthlib
opt-einsum==3.3.0
    # via tensorflow
pillow==8.1.2
    # via -r requirements.in
protobuf==3.15.5
    # via
    #   tensorboard
    #   tensorflow
pyasn1-modules==0.2.8
    # via google-auth
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pydantic==1.8.1
    # via fastapi
python-dotenv==0.15.0
    # via uvicorn
python-multipart==0.0.5
    # via -r requirements.in
pytz==2021.1
    # via django
pyyaml==5.4.1
    # via uvicorn
requests-oauthlib==1.3.0
    # via google-auth-oauthlib
requests==2.25.1
    # via
    #   requests-oauthlib
    #   tensorboard
rsa==4.7.2
    # via google-auth
six==1.15.0
    # via
    #   absl-py
    #   astunparse
    #   google-auth
    #   google-pasta
    #   grpcio
    #   h5py
    #   keras-preprocessing
    #   protobuf
    #   python-multipart
    #   tensorboard
    #   tensorflow
sqlparse==0.4.1
    # via django
starlette==0.13.6
    # via fastapi
tensorboard-plugin-wit==1.8.0
    # via tensorboard
tensorboard==2.4.1
    # via tensorflow
tensorflow-estimator==2.4.0
    # via tensorflow
tensorflow==2.4.1
    # via -r requirements.in
termcolor==1.1.0
    # via tensorflow
typing-extensions==3.7.4.3
    # via
    #   pydantic
    #   tensorflow
urllib3==1.26.3
    # via requests
uvicorn[standard]==0.13.4
    # via -r requirements.in
uvloop==0.15.2
    # via uvicorn
watchgod==0.7
    # via uvicorn
websockets==8.1
    # via uvicorn
werkzeug==1.0.1
    # via tensorboard
wheel==0.36.2
    # via
    #   astunparse
    #   tensorboard
    #   tensorflow
wrapt==1.12.1
    # via tensorflow

# The following packages are considered to be unsafe in a requirements file:
# setuptools
